+++ packages/libraries/pwned/test/matching.spec.ts
@@ -1,7 +1,7 @@
 import { zxcvbnOptions } from '@zxcvbn-ts/core/src'
 import matchPwnedFactory from '../src'
 
-const fetch = jest.fn(() => ({
+const fetch = jest.fn(async () => ({
   text() {
     return `008A205652858375D71117A63004CC75167:5\r\n3EA386688A0147AB736AABCEDE496610382:244`
   },